made all fixes
