+++ src/test/java/com/rakovets/course/java/core/practice/date_and_time/DateAndTimeDemo.java
@@ -3,9 +3,11 @@ package com.rakovets.course.java.core.practice.date_and_time;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 
-public class DateWrapperDemo {
+public class DateAndTimeDemo {
     public static void main(String[] args) {
-        DateWrapper dateWrapper = new DateWrapper();
+        int numberOfDays = 45;
+        TemporalAdjusterImplementationOne temporalAdjusterImplementationOne = new TemporalAdjusterImplementationOne();
+        TemporalAdjusterImplementationTwo  temporalAdjusterImplementationTwo = new TemporalAdjusterImplementationTwo(numberOfDays);
         DateTimeFormatter dateTimeFormatter = DateTimeFormatter.ofPattern("M d, yyyy");
         LocalDate dateOne = LocalDate.of(2001, 7,3);
         LocalDate dateTwo = LocalDate.now();
@@ -14,7 +16,7 @@ public class DateWrapperDemo {
         System.out.println(DateWrapper.formatDateToString(dateOne,dateTimeFormatter));
         System.out.println(DateWrapper.stringFormatToLocalDate("2007-12-03"));
         System.out.println(DateWrapper.gapeBetweenDatas(dateTwo,dateOne));
-        System.out.println(dateWrapper.adjustInto(dateTwo,45));
-        System.out.println(dateWrapper.adjustInto(dateOne));
+        System.out.println(temporalAdjusterImplementationTwo.adjustInto(dateTwo));
+        System.out.println(temporalAdjusterImplementationOne.adjustInto(dateOne));
     }
 }
